solute Day11 Part Two

+++ src/de/plikoener/adventToCode2022/Day11/Item.java
@@ -1,17 +1,17 @@
 package de.plikoener.adventToCode2022.Day11;
 
 public class Item {
-    private int value;
+    private long value;
 
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 
-    public Item(int value) {
+    public Item(long value) {
         this.value = value;
     }
 
-    public void setValue(int value) {
+    public void setValue(long value) {
         this.value = value;
     }
 }
